requirements.txt
streamlit>=1.32
langgraph>=0.0.40
langchain>=0.1.0

google-generativeai>=0.3.2

numpy>=1.24
pandas>=2.0
matplotlib>=3.7

QuantLib>=1.32

scipy>=1.10

pyportfolioopt>=1.5
yfinance>=0.2.18
plotly>=5.18.0
